+++ packages/market-data/src/index.js
@@ -1,2 +1,5 @@
 export { MarketDataProvider } from "./MarketDataProvider.js";
+export { StooqMarketDataProvider, parseStooqCsv } from "./StooqMarketDataProvider.js";
 export { YahooMarketDataProvider } from "./YahooMarketDataProvider.js";
+export { FallbackMarketDataProvider } from "./FallbackMarketDataProvider.js";
+export { createDefaultMarketDataProvider } from "./factory.js";